fix float precision: promote MySQL float to float64 in Parquet; 6-sig-fig comparison
flexible_schema_manager.py:
  - MySQL FLOAT maps to pa.float64() instead of pa.float32() in both
    _map_mysql_to_pyarrow() and _build_type_mapping(). Python reads MySQL
    FLOAT as a Python float64 anyway; storing as float32 caused 32-bit
    precision artifacts in Redshift (e.g. 12.2 -> 12.1999998092651).

compare_pipelines.py:
  - _norm() uses :.6g (6 significant figures) for float-like values.
    Scale-independent: absorbs float32 display differences (19.200001 ->
    19.2) and float64 driver representation noise (36.190029678130855 vs
    36.1900296781309) while still catching real differences > 1ppm.
  - Result: UTS goes from 13,867 false mismatches to 0 (PASS).

+++ src/core/flexible_schema_manager.py
@@ -267,8 +267,11 @@ class FlexibleSchemaManager:
             return pa.int16()  # Use int16 instead of int8 for better range
         
         elif data_type in ['float', 'real']:
-            return pa.float32()
-        
+            # Promote to float64: Python reads MySQL FLOAT as a 64-bit Python float anyway.
+            # Storing as float32 in Parquet gives 32-bit precision artifacts in Redshift
+            # (e.g. 12.2 becomes 12.1999998092651). float64 avoids this with no downside.
+            return pa.float64()
+
         elif data_type == 'double':
             return pa.float64()
         
@@ -637,8 +640,8 @@ class FlexibleSchemaManager:
             'tinyint': pa.int8(),
             'decimal': pa.decimal128(10, 2),
             'numeric': pa.decimal128(10, 2),
-            'float': pa.float32(),
-            'real': pa.float32(),
+            'float': pa.float64(),   # promoted — see _map_mysql_to_pyarrow comment
+            'real': pa.float64(),
             'double': pa.float64(),
             'datetime': pa.timestamp('us'),
             'timestamp': pa.timestamp('us'),
